% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datum2df.R
\name{datum2df}
\alias{datum2df}
\title{datum2df}
\usage{
datum2df(datum, vars = c("gpp"), pfts = c(2, 3, 4, 17), name = "ref")
}
\arguments{
\item{datum}{datum}

\item{vars}{vars}

\item{pfts}{pfts}

\item{stderr}{should stderr be returned as well.}
}
\value{
df
}
\description{
datum2df
}
\author{
FM
}
