% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2ncdf_datum.R
\name{model2ncdf_datum}
\alias{model2ncdf_datum}
\title{Write PEcAn standard output files}
\usage{
model2ncdf_datum(
  datum,
  sitelat = -9.15,
  sitelon = -80,
  start_date = "2004/01/01"
)
}
\description{
Write PEcAn standard output files
}
